refactor(orlando): decompor OrlandoApp.tsx em 6 subcomponentes
Chunk 6 do plano de decomposição (issue #481): extrai OrlandoHero,
OrlandoFeatures, OrlandoParksGallery, OrlandoItinerary e OrlandoFooter
de OrlandoApp.tsx (838 → 36 linhas), zero mudança de comportamento.
